flaot

+++ cifar_train.py
@@ -362,7 +362,7 @@ def validate(val_loader, model, criterion, epoch, args, log=None, tf_writer=None
                 # output_randn = torch.cat(randn_list,dim=1)
                 output_list = []
                 # output_randn = torch.randn(val_batch_size,device='cuda')
-                output_randn = torch.tensor([i for i in range(val_batch_size)],device='cuda',dtype='float')
+                output_randn = torch.tensor([i for i in range(val_batch_size)],device='cuda').float()
                 for c in range(num_classes):
                     output_list.append(output_randn.sort()[0][output[:,c].sort()[1]].view(val_batch_size,1))
                 output = torch.cat(output_list,dim=1)
